Move public image builds to CircleCI
GitHub actions have a disk space limit of 14G. Our images
do not fit into that. CircleCI does not have this limitation,
so let's switch to that.

While deployment that requires access to credentials will
happen with space telescope internal Jenkins, CircleCI can
be used to run external *untrusted* code. Since anyone can
make a PR, it'll be more secure to run this CI in an
external system that has no access to our AWS infrastructure